Refactor spec package
