+++ newrelic-integration-e2e/internal/agent/integration.go
@@ -1,6 +1,8 @@
 package agent
 
-import e2e "github.com/newrelic/newrelic-integration-e2e-action/newrelic-integration-e2e/internal"
+import (
+	"github.com/newrelic/newrelic-integration-e2e-action/newrelic-integration-e2e/internal/spec"
+)
 
 type integration struct {
 	Name   string                 `yaml:"name"`
@@ -10,7 +12,7 @@ type integrationList struct {
 	Integrations []integration `yaml:"integrations"`
 }
 
-func getIntegrationList(integrations []e2e.Integration) *integrationList {
+func getIntegrationList(integrations []spec.Integration) *integrationList {
 	out := &integrationList{
 		Integrations: make([]integration, len(integrations)),
 	}
