+++ newrelic-integration-e2e/internal/spec/definition.go
@@ -1,4 +1,4 @@
-package e2e
+package spec
 
 import yaml "gopkg.in/yaml.v3"
