+++ newrelic-integration-e2e/internal/spec/definition_test.go
@@ -1,4 +1,4 @@
-package e2e
+package spec
 
 import (
 	"testing"
